+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* **New** - Health Check for cron schedule
+
 ### 1.1.7: June 13th, 2019
 * **Tweak** - Deregistrer Moment from WP core
 * **Tweak** - Update Moment from 2.22.2 to 2.24